+++ src/ingredient/ingredient.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
-import { IIngredient } from 'src/common/interfaces/ingredient.interface';
+import { IIngredient } from './interface/ingredient.interface';
 import { INGREDIENT } from 'src/common/models/models';
 import { IngredientDTO } from './dto/ingrefient.dto';
