+++ src/product/interface/product.interface.ts
@@ -1,5 +1,3 @@
-import { IIngredient } from "./ingredient.interface";
-
 export interface IProduct extends Document{
   name: string;
   price: number;
